ai-training/common/training_utils.py
from __future__ import annotations

from pathlib import Path
from typing import Any, Dict

import yaml


def load_yaml(path: str) -> Dict[str, Any]:
    return yaml.safe_load(Path(path).read_text())


def ensure_output_dir(name: str) -> Path:
    out = Path("outputs") / name
    out.mkdir(parents=True, exist_ok=True)
    return out
