ai-training/requirements.txt
ultralytics==8.3.25
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
onnx==1.17.0
onnxruntime==1.19.2
PyYAML==6.0.2
numpy==2.1.2
